+++ pkg/server/commands/check_command.go
@@ -4,6 +4,7 @@ import (
 	"context"
 	"errors"
 	"math"
+	"sync"
 	"time"
 
 	"golang.org/x/sync/singleflight"
@@ -34,6 +35,7 @@ type CheckQuery struct {
 	datastore              storage.RelationshipTupleReader
 	cacheController        cachecontroller.CacheController
 	cacheSingleflightGroup *singleflight.Group
+	cacheWaitGroup         *sync.WaitGroup
 	checkCache             storage.InMemoryCache[any]
 	checkCacheTTL          time.Duration
 	maxCheckCacheSize      uint32
@@ -71,12 +73,22 @@ func WithCheckCommandLogger(l logger.Logger) CheckQueryOption {
 }
 
 // TODO can we make this better? There are too many caching flags.
-func WithCheckCommandCache(serverCtx context.Context, ctrl cachecontroller.CacheController, shouldCache bool, sf *singleflight.Group, cc storage.InMemoryCache[any], m uint32, ttl time.Duration) CheckQueryOption {
+func WithCheckCommandCache(
+	serverCtx context.Context,
+	ctrl cachecontroller.CacheController,
+	shouldCache bool,
+	sf *singleflight.Group,
+	cc storage.InMemoryCache[any],
+	wg *sync.WaitGroup,
+	m uint32,
+	ttl time.Duration,
+) CheckQueryOption {
 	return func(c *CheckQuery) {
 		c.cacheController = ctrl
 		c.shouldCacheIterators = shouldCache
 		c.serverCtx = serverCtx
 		c.cacheSingleflightGroup = sf
+		c.cacheWaitGroup = wg
 		c.checkCache = cc
 		c.maxCheckCacheSize = m
 		c.checkCacheTTL = ttl
@@ -128,7 +140,18 @@ func (c *CheckQuery) Execute(ctx context.Context, params *CheckCommandParams) (*
 		LastCacheInvalidationTime: cacheInvalidationTime,
 	}
 
-	requestDatastore := storagewrappers.NewRequestStorageWrapperForCheckAPI(c.serverCtx, c.datastore, params.ContextualTuples.GetTupleKeys(), c.maxConcurrentReads, c.shouldCacheIterators, c.cacheSingleflightGroup, c.checkCache, c.maxCheckCacheSize, c.checkCacheTTL)
+	requestDatastore := storagewrappers.NewRequestStorageWrapperForCheckAPI(
+		c.serverCtx,
+		c.datastore,
+		params.ContextualTuples.GetTupleKeys(),
+		c.maxConcurrentReads,
+		c.shouldCacheIterators,
+		c.cacheSingleflightGroup,
+		c.cacheWaitGroup,
+		c.checkCache,
+		c.maxCheckCacheSize,
+		c.checkCacheTTL,
+	)
 
 	ctx = typesystem.ContextWithTypesystem(ctx, c.typesys)
 	ctx = storage.ContextWithRelationshipTupleReader(ctx, requestDatastore)